Перенес модель пользователя в приложение api. Удалил приложение users.
