+++ api_yamdb/api/admin.py
@@ -1,3 +1,20 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import User
+
+
+class UserAdmin(admin.ModelAdmin):
+    list_display = (
+        'first_name',
+        'last_name',
+        'username',
+        'role',
+        'email',
+        'bio'
+        )
+    list_editable = ('role',)
+    search_fields = ('username',)
+    empty_value_display = '-пусто-'
+
+
+admin.site.register(User, UserAdmin)
